Logger&Migration

+++ src/main/java/com/sydoruk/service/CarHibernateService.java
@@ -2,6 +2,8 @@ package com.sydoruk.service;
 
 import com.sydoruk.model.*;
 import com.sydoruk.repository.CarHibernateRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,6 +13,7 @@ public class CarHibernateService {
 
     private final CarHibernateRepository repository;
     private final Random random = new Random();
+    final static Logger logger = LoggerFactory.getLogger(CarHibernateService.class);
     private final String[] types = {"GASOLINE","ELECTRIC","DIESEL"};
 
     public CarHibernateService(final  CarHibernateRepository repository) {
@@ -23,12 +26,18 @@ public class CarHibernateService {
             car = new PassengerCarHibernate();
         } else if (type == Type.TRUCK) {
             car = new TruckHibernate();
+        } else {
+            logger.error("Incorrect car`s type: {}", type);
         }
-        car.setEngine(createEngine());
-        car.setManufacturer(Manufacturer.randomManufacturer());
-        car.setColor(Color.randomColor());
-        car.setPrice(random.nextInt(1000, 100001));
-        car.setCount(random.nextInt(1, 6));
+        if (car != null) {
+            car.setEngine(createEngine());
+            car.setManufacturer(Manufacturer.randomManufacturer());
+            car.setColor(Color.randomColor());
+            car.setPrice(random.nextInt(1000, 100001));
+            car.setCount(random.nextInt(1, 6));
+        } else {
+            logger.warn("car is not created");
+        }
         return car;
     }
 
@@ -36,25 +45,41 @@ public class CarHibernateService {
         EngineHibernate engine = new EngineHibernate();
         engine.setPower(random.nextInt(0,1001));
         engine.setType(types[random.nextInt(types.length)]);
+        logger.info("Create {} engine {} power", engine.getType(), engine.getPower());
         return engine;
     }
 
     public CarHibernate createAndSave() {
         final CarHibernate car = createCar(Type.randomType());
-        repository.save(car);
+        if(car != null){
+            repository.save(car);
+        } else {
+            logger.error("Car cannot be saved");
+        }
         return car;
     }
 
     public Optional<CarHibernate> getById(final String id) {
+        logger.info("Search for car with id {}", id);
         return repository.getById(id);
     }
 
     public List<CarHibernate> getAll() {
-        return repository.getAll();
+        List<CarHibernate> cars = repository.getAll();
+        if (cars.isEmpty()){
+            logger.info("There are no cars in DB");
+        }
+        return cars;
     }
 
     public void delete(final String id) {
-        repository.delete(id);
+        logger.info("Delete car with id {}", id);
+        if (!(id == null || id.isEmpty())) {
+            repository.delete(id);
+        } else {
+            logger.warn("Incorrect id for deleting car");
+        }
+
     }
 
     public void printCar(CarHibernate car) {
